requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Deprecated==1.2.13
ecdsa==0.17.0
fastapi==0.70.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
jose==1.0.0
jwcrypto==1.0
jwt==1.3.1
oauth2client==4.1.3
passlib==1.7.4
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.8.2
pyparsing==3.0.6
Pyrebase4==4.5.0
python-jose==3.3.0
python-jwt==3.3.1
python-multipart==0.0.5
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
wrapt==1.13.3